requirements.txt
requests>=2.31.0
httpx>=0.24.1
beautifulsoup4>=4.12.2
markdown>=3.5.1
python-dotenv>=1.0.0
html5lib>=1.1
colorama>=0.4.6
tqdm>=4.66.1
tenacity>=8.2.3
google-generativeai>=0.3.2
lxml>=4.9.3
aiohttp>=3.9.1
bs4>=0.0.1
